Solution to Check Month Days

+++ Check Month Days@@6788dd63490637cfc44e1baa/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    int x;
-    scanf("%d", &x);
+    double  x;
+    scanf("%lf", &x);
     switch(x){
         case('1','3','5','7','10','12'):
                 printf("31");
